refactor: change the pos file structure and add getMdToEditorPos helper
