+++ apps/editor/src/markdown/marks/code.ts
@@ -3,7 +3,7 @@ import { TextSelection } from 'prosemirror-state';
 import { Command } from 'prosemirror-commands';
 import { cls } from '@/utils/dom';
 import Mark from '@/spec/mark';
-import { interpolatePos } from './helper/pos';
+import { interpolatePos } from '../helper/pos';
 
 const reCode = /^(`).*([\s\S]*)\1$/m;
 const codeSyntax = '`';
@@ -51,7 +51,7 @@ export class Code extends Mark {
       if (reCode.test(textContent)) {
         tr = tr.delete(to - 1, to).delete(from, from + 1);
       } else {
-        tr = tr.insertText(codeSyntax, to, to).insertText(codeSyntax, from, from);
+        tr = tr.insertText(codeSyntax, to).insertText(codeSyntax, from);
         const selection = empty
           ? TextSelection.create(tr.doc, from + 1, from + 1)
           : TextSelection.create(tr.doc, from, to + 2);
