--- apps/editor/src/markdown/marks/helper/pos.ts
@@ -1,11 +1,0 @@
-import { AllSelection, Selection } from 'prosemirror-state';
-
-export function interpolatePos(selection: Selection) {
-  let { from, to } = selection;
-
-  if (selection instanceof AllSelection) {
-    from += 1;
-    to -= 1;
-  }
-  return [from, to];
-}
