+++ apps/editor/src/markdown/marks/strong.ts
@@ -3,7 +3,7 @@ import { TextSelection } from 'prosemirror-state';
 import { Command } from 'prosemirror-commands';
 import { cls } from '@/utils/dom';
 import Mark from '@/spec/mark';
-import { interpolatePos } from './helper/pos';
+import { interpolatePos } from '../helper/pos';
 
 const reStrong = /^(\*{2}|_{2}).*([\s\S]*)\1$/m;
 const strongSyntax = '**';
@@ -37,7 +37,7 @@ export class Strong extends Mark {
       if (reStrong.test(textContent)) {
         tr = tr.delete(to - 2, to).delete(from, from + 2);
       } else {
-        tr = tr.insertText(strongSyntax, to, to).insertText(strongSyntax, from, from);
+        tr = tr.insertText(strongSyntax, to).insertText(strongSyntax, from);
         const selection = empty
           ? TextSelection.create(tr.doc, from + 2, from + 2)
           : TextSelection.create(tr.doc, from, to + 4);
